+++ src/components/Fields/RichText/RichText.js
@@ -17,6 +17,7 @@ import { toggleMark } from './lib/actions';
 import serialize from './lib/serialize';
 import parse, { defaultValue } from './lib/parse';
 import { INLINE_DISALLOWED_ITEMS } from './lib/options';
+import blockquotes from './lib/blockquotes';
 import Element from './Element';
 import Leaf from './Leaf';
 import Toolbar from './Toolbar';
@@ -117,6 +118,8 @@ const RichText = ({
     [disallowedTypesWithDefaults.join()],
   );
 
+  const { toggleBlockquote } = blockquotes(editor);
+
   // Set starting state from prop value on start up
   useEffect(() => {
     parse(initialValue)
@@ -164,6 +167,12 @@ const RichText = ({
             onKeyDown={handleKeyDown}
           />
         </div>
+        <button
+          type="button"
+          onClick={toggleBlockquote}
+        >
+          toggle blockquote
+        </button>
       </RichTextContainer>
     </Slate>
   );
